Moved main main toolbar to a seperate folder
