+++ main.js
@@ -2,7 +2,6 @@ const {app, BrowserWindow, Menu, ipcMain, dialog, shell} = require('electron');
 const path = require('path');
 const {download} = require("electron-dl");
 const fs = require("fs");
-const trash = require('trash');
 const {autoUpdater} = require("electron-updater");
 var extract = require('extract-zip');
 var request = require('request');
@@ -14,8 +13,9 @@ let loginWindow;
 function createWindow () {
   mainWindow = new BrowserWindow({width: 800, height: 600});
   mainWindow.loadFile('index.html');
-
-  const mainMenu = Menu.buildFromTemplate(mainMenuTemplate);
+  
+  let mainMenuToolbar = require("./src/modules/MainMenuToolbar.js");
+  const mainMenu = Menu.buildFromTemplate(mainMenuToolbar);
   // Insert menu
   Menu.setApplicationMenu(mainMenu);
 
@@ -203,74 +203,4 @@ app.on('window-all-closed', function () {
 
 app.on('activate', function () {
     createWindow();
-});
-
-const mainMenuTemplate = [
-  {
-    label: "File",
-    submenu:[
-      {
-        label:'Login',
-        click(){
-          logWindow();
-        }
-      },
-      {
-        label:'Delete MultiMC folder',
-        click(){
-          let options  = {
-            buttons: ["Yes","Cancel"],
-            message: "Do you really want to delete MultiMC? This sends your MultiMC to the trash!"
-          };
-          var cancel = dialog.showMessageBox(options);
-          if(!cancel){
-            trash(path.join(app.getPath("userData"), "process")).then(function(){
-              mainWindow.reload();
-            });
-          }
-        }
-      },
-      {
-        label:'Launch MultiMC normally',
-        click(){
-          mainWindow.webContents.executeJavaScript('App.launchNoArgs()');
-        }
-      },
-      {
-        label:'Open process folder',
-        click(){
-          shell.openItem(path.join(app.getPath("userData"), "process"));
-        }
-      },
-      {
-        label:'Quit',
-        accelerator: process.platform == 'darwin' ? 'Command+Q' : 'Ctrl+Q',
-        click(){
-          app.quit();
-        }
-      }
-    ]
-  }
-];
-
-// If on a mac, add a empty object on the menu
-if(process.platform == 'darwin'){
-  mainMenuTemplate.unshift({});
-}
-
-// Add developer tools
-mainMenuTemplate.push({
-  label: 'Developer Tools',
-  submenu:[
-    {
-      label: "Toggle DevTools",
-      accelerator: process.platform == 'darwin' ? 'Command+I' : 'Ctrl+I',
-      click(item, focusedWindow){
-        focusedWindow.toggleDevTools();
-      }
-    },
-    {
-      role: 'reload'
-    }
-  ]
 });
